pyproject.toml
[project]
name = "url-shortener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioredis==2.0.1",
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "fastapi==0.104.1",
    "httpx==0.25.2",
    "psycopg2-binary==2.9.9",
    "pydantic[email]==2.5.0",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "redis==5.0.1",
    "sqlalchemy==2.0.23",
    "uvicorn[standard]==0.24.0",
]
